+++ app/src/models/modlog.lua
@@ -1,0 +1,8 @@
+--- Moderation log model
+-- @module models.modlog
+
+local Model = require("lapis.db.model").Model
+
+return Model:extend("modlog", {
+	timestamp = true,
+})